Introduces Regular Page Layouts
Regular page layout allows me to make posts (layout: regpage) that have
a full page URL with a header and footer. Project date, title and
description are inserted at the top, and there’s padding to make this
look like sth resembling a regular blog post, rather than a fully
controllable panel (planning to introduce fullpage layouts to grab the
entire panel for more control).

This allows me to do full pages in HTML using JS and other stuff if I
want to, while still leveraging the post system. Or I can just write
them in markdown or whatever.

These posts still appear in the home page’s portfolio grid, so they
ought to have an image and an alt image to appear there. The popup
links to the page. The post description is used there.